refactor(ms-user): Change user ID data type from Integer to Long
In this commit, the data type of the user ID has been changed from Integer to Long to accommodate larger ID values. This change required updates in the User model, UserResponseDTO, UserRepository, UserService, and UserController.

+++ backend/ms-user/src/main/java/com/passportal/msuser/controller/UserController.java
@@ -54,7 +54,7 @@ public class UserController {
 
     // TO DO: Authorizations to delete user.
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable Integer id) {
+    public ResponseEntity<?> delete(@PathVariable Long id) {
         try {
             userService.deleteById(id);
             return ResponseEntity.ok("User successfully deleted");
@@ -64,7 +64,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getById(@PathVariable Integer id) {
+    public ResponseEntity<?> getById(@PathVariable Long id) {
         try {
             UserResponseDTO userResponseDTO = userService.getById(id);
             return ResponseEntity.ok(userResponseDTO);
